Separate EmailForm from ResetPasswordMain component
