+++ src/login/components/LoginApp/ResetPassword/ResetPasswordMain.tsx
@@ -1,69 +1,17 @@
 import React, { useEffect } from "react";
 import InjectIntl from "../../../translation/InjectIntl_HOC_factory";
-import { connect } from "react-redux";
 import { useAppDispatch, useAppSelector } from "../../../app_init/hooks";
 import resetPasswordSlice from "../../../redux/slices/resetPasswordSlice";
-import { Field, reduxForm, InjectedFormProps } from "redux-form";
-import Form from "reactstrap/lib/Form";
-import CustomInput from "../../Inputs/CustomInput";
-import EduIDButton from "../../../../components/EduIDButton";
-import { validate } from "../../../app_utils/validation/validateEmail";
+import { InjectedFormProps } from "redux-form";
 import PropTypes from "prop-types";
 import { clearCountdown, setLocalStorage } from "./CountDownTimer";
 import Splash from "../../../../containers/Splash";
+import EmailForm, { EmailFormData, EmailFormProps } from "./EmailForm";
 
 export const LOCAL_STORAGE_PERSISTED_EMAIL = "email";
 
-interface EmailFormData {
-  email?: string;
-}
-interface EmailFormProps {
-  requestEmailLink: (event: React.FormEvent<HTMLFormElement>) => void;
-  translate(msg: string): string;
-  request_in_progress: boolean;
-  invalid: boolean;
-}
-
-const EmailForm: React.FC<EmailFormProps & InjectedFormProps<EmailFormData, EmailFormProps>> = (
-  props: EmailFormProps
-): JSX.Element => {
-  return (
-    <Form id="reset-password-form" role="form" onSubmit={props.requestEmailLink}>
-      <Field
-        type="email"
-        name="email"
-        label={props.translate("profile.email_display_title")}
-        componentClass="input"
-        id="email-input"
-        component={CustomInput}
-        placeholder="name@example.com"
-        required={true}
-        helpBlock={props.translate("emails.input_help_text")}
-      />
-      <EduIDButton
-        className="settings-button"
-        id="reset-password-button"
-        disabled={props.invalid || props.request_in_progress}
-        // onClick={sendLink}
-      >
-        {props.translate("resetpw.send-link")}
-      </EduIDButton>
-    </Form>
-  );
-};
-
-const AllEmailForm = reduxForm<EmailFormData, EmailFormProps>({
-  form: "reset-pass-email-form",
-  validate,
-  touchOnChange: true,
-  enableReinitialize: true,
-  destroyOnUnmount: false,
-})(EmailForm);
-
-connect(() => ({}))(AllEmailForm);
-
 type ErrorType = {
-  msg: unknown[];
+  msg: string[];
 };
 
 interface ResetPasswordMainProps {
@@ -104,7 +52,10 @@ function ResetPasswordMain(
 
   return (
     <>
-      {errors && typeof errors[0] && (errors[0] as ErrorType).msg.includes("phone-code") && <Splash />}
+      {errors &&
+        typeof errors[0] !== undefined &&
+        (errors[0] as ErrorType).msg !== undefined &&
+        (errors[0] as ErrorType).msg.includes("phone-code") && <Splash />}
       <p className="heading">{props.translate("resetpw.heading-add-email")}</p>
       <EmailForm requestEmailLink={requestEmailLink} {...props} request_in_progress={request_in_progress} />
       <div className={loginRef ? `return-login-link` : `return-login-link disabled`}>
